"Added removeCage function and delete button to cages"
Model used: gpt-4-0613

me> Ok, please add a "X" button in each cage control.  If the X is clicked, that cage should be removed.
jr> I've added an "X" button to each cage control. When this button is clicked, the corresponding cage will be removed. Please refresh your browser to see the changes.