"Refactored cell mode toggle logic and updated cell rendering to support new data structure"
Model used: gpt-4-0613

me> Hmmm, same thing still happens.
jr> I've updated the `toggleMode` function in your `index.html` file. Now, when you click on a digit, only that digit's state should change, not the states of all the digits in the cell. Please check to see if the issue is resolved.