"Added function to generate and display possible values for each cage in the game board"
Model used: gpt-4-0613

me> Please update index.html so that the possibility cages contain all the possible combinations of digits that sastify the set of constraints imposed by its cage. For example, if the there was a 3-cell cage that had {1,2} as the only available digits for the first cell, {3,4} as the only digits available for the second cell, and 5 as a required digit for the third cell, then the possibilities printed out would be: [ [ 1,3,5 ], [ 2,3,5 ], [ 1,4,5 ], [ 2,4,5 ] ].
jr> I've updated the `index.html` file to include the functionality you requested. Now, each cage will display all possible combinations of digits that satisfy the set of constraints imposed by its cage. The possibilities are calculated based on the available and required digits for each cell in the cage.