"Moved cage possibilities section inside conditional rendering block"
Model used: gpt-4-0613

me> Please update index.html and move the display of the possibiliy cages to be below the rendered board (and only show up when it does).
jr> I've updated the `index.html` file to move the display of the possibility cages below the rendered board. They will now only show up when the board is rendered.