"Refactor condition check in generate function for code simplification"
Model used: gpt-4-0613

me> Oh, yeah, that's not right.  The idea is this: Let's say I had created a three cell cage - in the first cell I marked 1 as required.  In the second cell I mark every digit except 2 as forbidden, in the third cell I mark every cell except 3 and 4 as forbidden.  In this case, the possibilities should show [1,2,3] and [1,2,4] because the 1 is requried in the first cell, everything but 2 is forbidden in the second cell, and both 3 and 4 are available for the last cell.  Does that make sense?
jr> I've updated the `generateCagePossibilities` function in your `index.html` file. Now, it should correctly generate possibilities where each value is either 'required' or 'available', but not 'forbidden'. This should allow the possibilities list to display the correct options based on your cage configurations.