"Updated cell mode condition in index.html to include 'required' and 'available' states"
Model used: gpt-4-0613

me> In index.html, the cage possibilities do not appropriately taken into account the required and forbidden numbers.
jr> I've updated the `generateCagePossibilities` function in the `index.html` file. Now, it will consider the required values as well. If a value is required, it will be included in every possibility. If a value is forbidden, it will not be included in any possibility. If a value is available, it can be included in some possibilities.