"Added max-width property to main container in index.html"
Model used: gpt-4-0613

me> Can you take a look at index.html and constrain the initial board on the right to be a max of say 500px wide?
jr> I've updated the `index.html` file to constrain the initial board on the right to be a maximum of 500px wide. This was done by adding a `max-width: 500px;` property to the `.board-container` class in the CSS.