"Added flex display style to .row class in index.html"
Model used: gpt-4-0613

me> In index.html, previously we tried to arrange the numbers in the rendered board to be in rows of 1,2,3 and 4,5,6 and 7,8,9, but instead the numbers are just all displayed vertically from 1-9.  Can you see if you can figure out how to fix that, please?
jr> I've updated the `index.html` file to fix the issue. The numbers should now be displayed in rows of 1,2,3 and 4,5,6 and 7,8,9 as expected.